add initial guess and get working
