+++ src/freqfit/guess.py
@@ -1,6 +1,6 @@
 """
 Abstract base class for initial guesses. Needs to take an Experiment and return a guess 
-for each fit parameter.
+for each fit parameter as a dict.
 """
 
 from abc import ABC, abstractmethod
